pyproject.toml
[tool.poetry]
name = "rag-doc-splitter"
version = "0.1.0"
description = ""
authors = ["Chaz Wilder <hello@chazwilder.io>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.7.2"
langchain = "^0.2.0"
llama-index = "^0.10.38"
pydantic = "^2.7.1"
openai = "^1.30.1"
semantic-text-splitter = "^0.13.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
